+++ changelog.md
@@ -4,6 +4,7 @@
 - Improved the toggling of the notification-based features
 - Added the "display exact vote counts on poll answers" feature
 - Added the "show hidden NSFW posts in the timeline" feature
+- Added the "messaging scale" feature
 
 ## v4.4.6
 - Changed the custom notification labels to use theme colours